redesign API around geometry functions
